ДЗ #7 Spring Data JPA Repositories

+++ bookeeper/src/main/java/com/etn319/service/impl/AuthorServiceImpl.java
@@ -1,13 +1,14 @@
 package com.etn319.service.impl;
 
-import com.etn319.dao.DaoLayerException;
-import com.etn319.dao.api.AuthorDao;
+import com.etn319.dao.AuthorRepository;
 import com.etn319.model.Author;
 import com.etn319.service.CacheHolder;
+import com.etn319.service.EntityNotFoundException;
 import com.etn319.service.ServiceLayerException;
 import com.etn319.service.api.AuthorService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -18,7 +19,7 @@ import java.util.Optional;
 @RequiredArgsConstructor
 @Service
 public class AuthorServiceImpl implements AuthorService {
-    private final AuthorDao dao;
+    private final AuthorRepository dao;
     private final CacheHolder cache;
 
     @Override
@@ -28,14 +29,14 @@ public class AuthorServiceImpl implements AuthorService {
 
     @Override
     public Optional<Author> getById(long id) {
-        Optional<Author> author = dao.getById(id);
+        Optional<Author> author = dao.findById(id);
         author.ifPresent(cache::setAuthor);
         return author;
     }
 
     @Override
     public List<Author> getAll() {
-        return dao.getAll();
+        return dao.findAll();
     }
 
     @Override
@@ -47,7 +48,7 @@ public class AuthorServiceImpl implements AuthorService {
             Author saved = dao.save(author);
             clearCache();
             return saved;
-        } catch (DaoLayerException e) {
+        } catch (DataAccessException e) {
             throw new ServiceLayerException(e);
         }
     }
@@ -55,9 +56,13 @@ public class AuthorServiceImpl implements AuthorService {
     @Override
     @Transactional
     public void deleteById(long id) {
+        if (!dao.existsById(id)) {
+            throw new EntityNotFoundException();
+        }
+
         try {
             dao.deleteById(id);
-        } catch (DaoLayerException e) {
+        } catch (DataAccessException e) {
             throw new ServiceLayerException(e);
         }
     }
