+++ bookeeper/src/main/java/com/etn319/service/impl/GenreServiceImpl.java
@@ -1,13 +1,14 @@
 package com.etn319.service.impl;
 
-import com.etn319.dao.DaoLayerException;
-import com.etn319.dao.api.GenreDao;
+import com.etn319.dao.GenreRepository;
 import com.etn319.model.Genre;
 import com.etn319.service.CacheHolder;
+import com.etn319.service.EntityNotFoundException;
 import com.etn319.service.ServiceLayerException;
 import com.etn319.service.api.GenreService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -18,7 +19,7 @@ import java.util.Optional;
 @Slf4j
 @RequiredArgsConstructor
 public class GenreServiceImpl implements GenreService {
-    private final GenreDao dao;
+    private final GenreRepository dao;
     private final CacheHolder cache;
 
     @Override
@@ -28,14 +29,14 @@ public class GenreServiceImpl implements GenreService {
 
     @Override
     public Optional<Genre> getById(long id) {
-        Optional<Genre> genre = dao.getById(id);
+        Optional<Genre> genre = dao.findById(id);
         genre.ifPresent(cache::setGenre);
         return genre;
     }
 
     @Override
     public List<Genre> getAll() {
-        return dao.getAll();
+        return dao.findAll();
     }
 
     @Override
@@ -46,7 +47,7 @@ public class GenreServiceImpl implements GenreService {
             Genre saved = dao.save(genre);
             clearCache();
             return saved;
-        } catch (DaoLayerException e) {
+        } catch (DataAccessException e) {
             throw new ServiceLayerException(e);
         }
     }
@@ -54,9 +55,13 @@ public class GenreServiceImpl implements GenreService {
     @Override
     @Transactional
     public void deleteById(long id) {
+        if (!dao.existsById(id)) {
+            throw new EntityNotFoundException();
+        }
+
         try {
             dao.deleteById(id);
-        } catch (DaoLayerException e) {
+        } catch (DataAccessException e) {
             throw new ServiceLayerException(e);
         }
     }
